trying to restart engine using another script but having trouble to get both scripts running in parallel
